Adding the simplified bioreactor simulation and other minor things.
The bioreactor for glycerol production was modeled in Aspen in the
simplest form possible (documentation for the process is in the
corresponding .org file). Other changes include minor calculation
changes in some excel files and some additions to the bioreactor
stoichiometry which mention what the simplest form of the bioreaction
that was simulated is.