Added files for the complete bioreactor simulation
In the /Aspen folder are the actual files of the simulations and a
file documenting the process. Supplementary to those is
biomass_modeling_aspen in the top level which is a file explaining how
I modeled the biomass in Aspen. Lastly in the /Calculations folder are
some matlab files I used for the simulation data.