Making a correction on the bioreactor simulation.
For the bioreactor simulation, there was a parameter that had not been
included in the simulation, while it was important. That was how much
biomass was produced according to the bibliography. The simulation has
been corrected and documented (to an extent). It will be fully
documented in the next commit.