Simulated the combustion of lignin in Aspen
For this simulation, some bibliographic data was extacted for the
kinetics of the process and based on existing info for the
stoichiometry the combustion process was simulated. Then, this process
was integrated with steam explosion to see how much lignin we need to
heat the steam used in the steam explosion process. All said processes
work with no errors